add Instruction::CopyF64Imm32
This is an optimization for copying constant f64 values that can be losslessly encoded as f32 values to avoid having to allocate a function local constant value.

+++ crates/wasmi/src/engine/bytecode2/immediate.rs
@@ -213,7 +213,7 @@ impl From<u32> for Const32<u64> {
     }
 }
 
-impl From<f32> for Const32<f32> {
+impl From<f32> for Const32<f64> {
     fn from(value: f32) -> Self {
         Self::new(AnyConst32::from_f32(F32::from(value)))
     }
